Add meal actions file
