+++ frontend/src/actions/meal_actions.js
@@ -1,0 +1,16 @@
+import * as MealplanAPIUtil from "../util/spoonacular_api_util";
+
+export const RECEIVE_MEAL_INFO = "RECEIVE_MEAL_INFO";
+
+export const receiveMealInfo = (meal) => {
+  return {
+    type: RECEIVE_MEAL_INFO,
+    meal,
+  };
+};
+
+export const getMealInfo = (id) => (dispatch) => {
+  return MealplanAPIUtil.getMealInfo(id).then((res) => {
+    dispatch(receiveMealInfo(res.data));
+  });
+};
